requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-google-genai==1.0.1
langchain-community==0.0.10
langgraph==0.0.26
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
smtplib
email-validator==2.1.0
twilio==8.10.0
openpyxl==3.1.2
plotly==5.17.0
datetime
uuid
requests==2.31.0
pydantic==2.5.2
typing-extensions==4.8.0
